Treat rrules with no recognized frequency as errors.
In the most current released version of libical, these calls will return
without error.  Note in the second example the separator is incorrect.

    icalrecurrencetype_from_string("FREQ=BANANA")
    icalrecurrencetype_from_string("FREQ=WEEKLY,BYDAY=MO")

However icalrecurrencetype_as_string() will not render this result back.
Instead it returns NULL, which causes the rrule output function to
segfault.

Since rrules are static in this extension, it makes sense to catch the
error in the input function.

It looks like a future release of libical will treat this as an error.
https://github.com/libical/libical/commit/5b99f67f6e7849dc679d53c591bfc87bef880ea9#diff-8161e632c3871bb32c8bc438d66b8c7fR536